refactor: extract TranscriptExportFormat.swift from TranscriptExporter.swift
Extracts the `TranscriptExportFormat` enum (text/markdown export
target metadata: title, fileExtension, contentType) from
TranscriptExporter.swift into its own file.

Byte-identical move of lines 5-35. New file imports Foundation and
UniformTypeIdentifiers (for UTType); TranscriptExporter.swift retains
AppKit + Foundation + UniformTypeIdentifiers.

TranscriptExporter.swift: 165 → 133 lines. Tests: 667.
