+++ Sources/BugNarrator/Services/TranscriptExportFormat.swift
@@ -1,0 +1,34 @@
+import Foundation
+import UniformTypeIdentifiers
+
+enum TranscriptExportFormat {
+    case text
+    case markdown
+
+    var title: String {
+        switch self {
+        case .text:
+            return "Export TXT"
+        case .markdown:
+            return "Export Markdown"
+        }
+    }
+
+    var fileExtension: String {
+        switch self {
+        case .text:
+            return "txt"
+        case .markdown:
+            return "md"
+        }
+    }
+
+    var contentType: UTType {
+        switch self {
+        case .text:
+            return .plainText
+        case .markdown:
+            return UTType(filenameExtension: "md") ?? .plainText
+        }
+    }
+}
